# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "algo-coach",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/cards/{slug}/runs": {
      "post": {
        "summary": "Started",
        "operationId": "started_api_cards__slug__runs_post",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cards/{slug}/recall/{template_id}": {
      "post": {
        "summary": "Recalled",
        "operationId": "recalled_api_cards__slug__recall__template_id__post",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reproduction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecallAttempt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sittings/{sitting_id}/submissions": {
      "post": {
        "summary": "Submission",
        "operationId": "submission_api_sittings__sitting_id__submissions_post",
        "parameters": [
          {
            "name": "sitting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sitting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Submission"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Submitted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sittings/{sitting_id}/pause": {
      "post": {
        "summary": "Paused",
        "operationId": "paused_api_sittings__sitting_id__pause_post",
        "parameters": [
          {
            "name": "sitting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sitting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Timed"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sittings/{sitting_id}/resume": {
      "post": {
        "summary": "Resumed",
        "operationId": "resumed_api_sittings__sitting_id__resume_post",
        "parameters": [
          {
            "name": "sitting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sitting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Timed"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sittings/{sitting_id}/end": {
      "post": {
        "summary": "Ended",
        "operationId": "ended_api_sittings__sitting_id__end_post",
        "parameters": [
          {
            "name": "sitting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sitting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Timed"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sittings/{sitting_id}/unclaimed": {
      "get": {
        "summary": "Unanswered",
        "operationId": "unanswered_api_sittings__sitting_id__unclaimed_get",
        "parameters": [
          {
            "name": "sitting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sitting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unclaimed"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/attempts/{attempt_id}/claims": {
      "post": {
        "summary": "Claimed",
        "operationId": "claimed_api_attempts__attempt_id__claims_post",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attempt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Claim"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptClaim"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Attempt": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "problem_id": {
            "type": "string",
            "title": "Problem Id"
          },
          "sitting_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sitting Id"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "type": "string",
            "format": "date-time",
            "title": "Finished At"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "time_to_solve_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time To Solve Sec"
          },
          "solved": {
            "type": "boolean",
            "title": "Solved"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "problem_id",
          "finished_at",
          "solved"
        ],
        "title": "Attempt"
      },
      "AttemptClaim": {
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "effort": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effort"
          },
          "prompt_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Hash"
          },
          "call_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Call Id"
          },
          "pin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pin"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "attempt_id": {
            "type": "string",
            "title": "Attempt Id"
          },
          "techniques": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Techniques"
          },
          "declined": {
            "type": "boolean",
            "title": "Declined",
            "default": false
          },
          "source": {
            "$ref": "#/components/schemas/ClaimSource"
          },
          "informed_by": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Informed By"
          },
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Confidence"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "attempt_id",
          "source"
        ],
        "title": "AttemptClaim",
        "description": "Which techniques an attempt used, as one writer claimed them."
      },
      "AttemptVerification": {
        "properties": {
          "cap_ms": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Cap Ms"
          },
          "runner": {
            "type": "string",
            "minLength": 1,
            "title": "Runner"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/CaseResult"
            },
            "type": "array",
            "title": "Results"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "attempt_id": {
            "type": "string",
            "title": "Attempt Id"
          }
        },
        "type": "object",
        "required": [
          "cap_ms",
          "runner",
          "id",
          "created_at",
          "attempt_id"
        ],
        "title": "AttemptVerification",
        "description": "Private to the user, where a solution's run is product data: `log.md`."
      },
...
      "CardRun": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "minLength": 1,
            "title": "User Id"
          },
          "card_id": {
            "type": "string",
            "minLength": 1,
            "title": "Card Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "probes": {
            "items": {
              "$ref": "#/components/schemas/Probe"
            },
            "type": "array",
            "title": "Probes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "card_id",
          "started_at"
        ],
        "title": "CardRun"
      },
      "CaseOutcome": {
        "type": "string",
        "enum": [
          "passed",
          "wrong",
          "timeout",
          "crashed"
        ],
        "title": "CaseOutcome"
      },
      "CaseResult": {
        "properties": {
          "case_id": {
            "type": "string",
            "minLength": 1,
            "title": "Case Id"
          },
          "outcome": {
            "$ref": "#/components/schemas/CaseOutcome"
          },
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Ms"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "outcome"
        ],
        "title": "CaseResult"
      },
      "Claim": {
        "properties": {
          "techniques": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Techniques",
            "default": []
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence"
          },
          "declined": {
            "type": "boolean",
            "title": "Declined",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "confidence"
        ],
        "title": "Claim"
      },
      "ClaimSource": {
        "type": "string",
        "enum": [
          "user",
          "classifier"
        ],
        "title": "ClaimSource"
      },
      "Confidence": {
        "type": "string",
        "enum": [
          "guess",
          "leaning",
          "sure"
        ],
        "title": "Confidence"
      },
      "Failure": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "outcome": {
            "$ref": "#/components/schemas/CaseOutcome"
          },
          "args": {
            "items": {},
            "type": "array",
            "title": "Args"
          },
          "expected": {
            "title": "Expected"
          },
          "returned": {
            "title": "Returned"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "outcome",
          "args",
          "expected",
          "returned",
          "error"
        ],
        "title": "Failure",
        "description": "The first case a submission failed, shown whole: `flows.md`."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Hint": {
        "type": "string",
        "enum": [
          "title",
          "notes",
          "form"
        ],
        "title": "Hint",
        "description": "What the trainer gives when memory fails, in the order it gives them.\nEach answers more of the question than the last."
      },
...
      "Pause": {
        "properties": {
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          },
          "until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until"
          }
        },
        "type": "object",
        "required": [
          "at"
        ],
        "title": "Pause"
      },
...
      "Probe": {
        "properties": {
          "problem_id": {
            "type": "string",
            "minLength": 1,
            "title": "Problem Id"
          },
          "assigned_at": {
            "type": "string",
            "format": "date-time",
            "title": "Assigned At"
          }
        },
        "type": "object",
        "required": [
          "problem_id",
          "assigned_at"
        ],
        "title": "Probe",
        "description": "One problem a run was given, testing whether the form is recognised\nunprompted. Drawn from outside the ladder, which teaches the form."
      },
...
      "RecallAttempt": {
        "properties": {
          "cap_ms": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Cap Ms"
          },
          "runner": {
            "type": "string",
            "minLength": 1,
            "title": "Runner"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/CaseResult"
            },
            "type": "array",
            "title": "Results"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "user_id": {
            "type": "string",
            "minLength": 1,
            "title": "User Id"
          },
          "card_id": {
            "type": "string",
            "minLength": 1,
            "title": "Card Id"
          },
          "template_id": {
            "type": "string",
            "minLength": 1,
            "title": "Template Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "hints": {
            "items": {
              "$ref": "#/components/schemas/Hint"
            },
            "type": "array",
            "title": "Hints"
          }
        },
        "type": "object",
        "required": [
          "cap_ms",
          "runner",
          "id",
          "created_at",
          "user_id",
          "card_id",
          "template_id",
          "code"
        ],
        "title": "RecallAttempt",
        "description": "A form typed into a blank file and run against the template's cases."
      },
...
      "Reproduction": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "hints": {
            "items": {
              "$ref": "#/components/schemas/Hint"
            },
            "type": "array",
            "title": "Hints",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "Reproduction",
        "description": "What the trainer sends: the form as typed, and the hints taken before\nit ran."
      },
...
      "Sitting": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "minLength": 1,
            "title": "User Id"
          },
          "problem_id": {
            "type": "string",
            "minLength": 1,
            "title": "Problem Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "last_active_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Active At"
          },
          "pauses": {
            "items": {
              "$ref": "#/components/schemas/Pause"
            },
            "type": "array",
            "title": "Pauses"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "problem_id",
          "started_at"
        ],
        "title": "Sitting"
      },
...
      "Submission": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "Submission"
      },
      "Submitted": {
        "properties": {
          "attempt": {
            "$ref": "#/components/schemas/Attempt"
          },
          "verification": {
            "$ref": "#/components/schemas/AttemptVerification"
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Failure"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "attempt",
          "verification",
          "failure"
        ],
        "title": "Submitted",
        "description": "The attempt a submission minted, and the per-case verdict behind its\n`solved`."
      },
...
      "Timed": {
        "properties": {
          "sitting": {
            "$ref": "#/components/schemas/Sitting"
          },
          "elapsed_sec": {
            "type": "number",
            "title": "Elapsed Sec"
          }
        },
        "type": "object",
        "required": [
          "sitting",
          "elapsed_sec"
        ],
        "title": "Timed",
        "description": "A sitting a pause, a resume or an end moved, with its elapsed time on the\nengine's clock."
      },
      "Unclaimed": {
        "properties": {
          "techniques": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Techniques"
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/Attempt"
            },
            "type": "array",
            "title": "Attempts"
          }
        },
        "type": "object",
        "required": [
          "techniques",
          "attempts"
        ],
        "title": "Unclaimed",
        "description": "The claim a sitting ends on: the attempts no claim answers yet, and the\nproblem's techniques they are answered over."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}